Use a filePath that is compatible with Jekyll
Changes the path of /.well-known/acme-challange/{key} to /letencrypt.html when --jekyll is used

Resolves 2 issues:
1. By default, Jekyll does not build any files or directories that start
with a '.'.  This causes the default path of starting with .well-known to not work when --jekyll flag is used.
2. Even if the --path option is used to point the {key} file to a
non-hidden directory, jekyll gets confused with the lack of a .html extension and doesn't render it properly